python2Packages.certifi: set knownVulnerabilities
Certifi is a python package providing a CA trust store. It's Python 2
support ended way back in 2019 and the code has been incompatible with
Python 2 ever since. We just recently noticed breakage, when tests were
enabled, so it was chosen to downgrade to the version that last
supported Python 2.7.

A CA store is a securitiy-critical affair and not being able to update
it any more makes this package a security risk.